// src/app/interfaces/i-anime.ts
export interface IAnime {
   

        airing: boolean;
        end_date: string;
        episodes: number;
        image_url: string;
        mal_id: string;
        members: number;
        rated: string;
        score: string;
        start_date: string;
        synopsis: string;
        title: string;
        type: string;
        url: string;

        orden: number;

}